requirements.txt
asgiref
attrs
autopep8
aws-psycopg2
awsebcli
bcrypt
blessed
boto3
botocore
cached-property
cement
certifi
cffi
chardet
colorama
cryptography
defusedxml
Django
django-allauth
django-ckeditor
django-js-asset
django-sslserver
django-storages
django-widget-tweaks
docker
docker-compose
dockerpty
docopt
docutils
fabric
future
idna
importlib-metadata
invoke
jmespath
jsonschema
mock
oauthlib
paramiko
pathlib2
pathspec
Pillow
prettytable
pyasn1
pycodestyle
pycparser
PyJWT
PyNaCl
pyrsistent
python-dateutil
python-decouple
python3-openid
pytz
PyYAML
relaxml
requests
requests-oauthlib
rsa
s3transfer
semantic-version
simplejson
six
sqlparse
termcolor
texttable
three
toml
typing-extensions
urllib3
wcwidth
websocket-client
zipp
